feat :: relocate package location
