+++ app/src/main/java/com/teamnk/kimiljung/feature/changepassword/ChangePasswordViewModel.kt
@@ -1,2 +1,2 @@
-package com.teamnk.kimiljung.feature.mypage
+package com.teamnk.kimiljung.feature.changepassword
