+++ app/src/main/java/com/teamnk/kimiljung/feature/mypage/MyPageFragment.kt
@@ -10,6 +10,8 @@ import androidx.lifecycle.ViewModelProvider
 import com.teamnk.kimiljung.R
 import com.teamnk.kimiljung.base.BaseFragment
 import com.teamnk.kimiljung.databinding.FragmentMypageBinding
+import com.teamnk.kimiljung.feature.changepassword.ChangePasswordActivity
+import com.teamnk.kimiljung.feature.changeuserinformation.ChangeUserInformationActivity
 import com.teamnk.kimiljung.feature.start.StartActivity
 import com.teamnk.kimiljung.util.showDialogWithDoubleButton
 import com.teamnk.kimiljung.util.showShortSnackBar
